feat: Add job-sink rock (#49)
Closes #48

This PR creates a rock for the job-sink images. The information is build to upstream image is located at https://github.com/knative/eventing/tree/knative-v1.16.1/cmd/jobsink.

This PR includes 2 small sanity checks to ensure that the final executable as well as the ca certificates exist in the correct path.

The upstream image uses ko to build.